developments with profile page and more
added the bulk of the profile page, including fetching user info. still
need to add getting and showing the users profile photo but that will be
the next logical step forward.
in addition, also changed some other layout info. i replaced a lot of
hardcoded values in the xml files with string tags in strings.xml to be
more compliant with the spec and to make changing things a little less
painful. also changed some layout arrangement, padding, and text colors
both to better fall in line with material design guidelines and to make
overall app ui consistency a thing which exists to a greater degree
(still some room to improve here but we can work on that as time
progresses)