made layout more dynamic
it used to be that the constraintlayouts in the schedule and assignment
recyclerview items would be fixed at a certain percent, and if you chose
a time format which made the textview extend past that length then too
bad. now the fragments calculate the min width to set all guidelines so
that width is both consistent and accommodating for user-set time/date
formats. the date/times chosen for calculating these widths were chosen
to represent the date which would precipiate the longest format string